requirements.txt
streamlit
sentence-transformers
pinecone-client
watchdog
nltk
PyMuPDF
requests
tqdm
